+++ guidata/tests/unit/test_validationmodes.py
@@ -7,11 +7,33 @@
 
 import os.path as osp
 
+import numpy as np
 import pytest
 
 import guidata.dataset as gds
 from guidata.config import ValidationMode, get_validation_mode, set_validation_mode
 from guidata.env import execenv
+
+
+def check_array(value: np.ndarray, raise_exception: bool = False) -> bool:
+    """Check if value is a valid 2D array of floats.
+
+    Args:
+        value: value to check
+        raise_exception: if True, raise an exception on invalid value
+
+    Returns:
+        True if value is valid, False otherwise
+    """
+    if (
+        not isinstance(value, np.ndarray)
+        or value.ndim != 2
+        or not np.issubdtype(value.dtype, np.floating)
+    ):
+        if raise_exception:
+            raise TypeError("Float array must be a 2D numpy array of floats")
+        return False
+    return True
 
 
 class Parameters(gds.DataSet):
@@ -20,6 +42,7 @@ class Parameters(gds.DataSet):
     fitem = gds.FloatItem("Float", min=1, max=250)
     iitem = gds.IntItem("Integer", max=20, nonzero=True)
     sitem = gds.StringItem("String", notempty=True)
+    aitem = gds.FloatArrayItem("Array", check_callback=check_array)
     fileopenitem = gds.FileOpenItem("File", ("py",))
     filesopenitem = gds.FilesOpenItem("Files", ("py",))
     filesaveitem = gds.FileSaveItem("Save file", ("py",))
@@ -30,6 +53,7 @@ VALID_DATA = {
     "fitem": 100.0,
     "iitem": 10,
     "sitem": "test",
+    "aitem": np.array([[1.0, 2.0], [3.0, 4.0]]),
     "fileopenitem": __file__,
     "filesopenitem": [
         __file__,
@@ -49,6 +73,10 @@ INVALID_DATA = {
         0,  # Zero not allowed
         "test",  # Not an integer
         23.2323,  # Not an integer
+    ),
+    "aitem": (
+        np.array([1.0, 2.0]),  # Not a 2D array
+        np.array([[1, 2], [3, 4]]),  # Not a float array
     ),
     "sitem": (
         "",  # Empty string not allowed
@@ -74,18 +102,45 @@ def test_default_validation_mode():
     execenv.print("OK")
 
 
+def __check_assigned_value_is_equal(assigned_value, expected_value):
+    """Check if the assigned value is correctly set"""
+    if isinstance(expected_value, np.ndarray):
+        # For arrays, we check if the value is set correctly
+        assert isinstance(assigned_value, np.ndarray)
+        assert assigned_value.shape == expected_value.shape
+        assert np.all(assigned_value == expected_value)
+    else:
+        # For other types, we check if the value is set correctly
+        assert assigned_value == expected_value
+
+
+def __check_assigned_value_is_not_equal(assigned_value, expected_value):
+    """Check if the assigned value is not equal to the real value"""
+    if isinstance(expected_value, np.ndarray):
+        # For arrays, we check if the value is set correctly
+        if isinstance(assigned_value, np.ndarray):
+            assert assigned_value.shape == expected_value.shape
+            assert not np.all(assigned_value == expected_value)
+        else:
+            assert assigned_value is None
+    else:
+        # For other types, we check if the value is set correctly
+        assert assigned_value != expected_value
+
+
 def test_valid_data():
     """Test valid data"""
     params = Parameters()
     execenv.print("Testing valid data: ", end="")
     for name, value in VALID_DATA.items():
         setattr(params, name, value)
-        assert getattr(params, name) == value
+        __check_assigned_value_is_equal(getattr(params, name), value)
     execenv.print("OK")
 
 
 def test_invalid_data_with_no_validation():
     """Test invalid data with validation disabled"""
+    old_mode = get_validation_mode()
     params = Parameters()
     set_validation_mode(ValidationMode.DISABLED)
     assert get_validation_mode() == ValidationMode.DISABLED
@@ -95,11 +150,13 @@ def test_invalid_data_with_no_validation():
             execenv.print(f"  Testing {name} with value: {value}")
             setattr(params, name, value)
             # No exception should be raised
-            assert getattr(params, name) == value
+            __check_assigned_value_is_equal(getattr(params, name), value)
+    set_validation_mode(old_mode)
 
 
 def test_invalid_data_with_enabled_validation():
     """Test invalid data with validation enabled"""
+    old_mode = get_validation_mode()
     params = Parameters()
     set_validation_mode(ValidationMode.ENABLED)
     assert get_validation_mode() == ValidationMode.ENABLED
@@ -111,11 +168,13 @@ def test_invalid_data_with_enabled_validation():
             with pytest.warns(gds.DataItemValidationWarning):
                 setattr(params, name, value)
             # The value should be set anyway
-            assert getattr(params, name) == value
+            __check_assigned_value_is_equal(getattr(params, name), value)
+    set_validation_mode(old_mode)
 
 
 def test_invalid_data_with_strict_validation():
     """Test invalid data with strict validation"""
+    old_mode = get_validation_mode()
     params = Parameters()
     set_validation_mode(ValidationMode.STRICT)
     assert get_validation_mode() == ValidationMode.STRICT
@@ -127,7 +186,8 @@ def test_invalid_data_with_strict_validation():
             with pytest.raises(gds.DataItemValidationError):
                 setattr(params, name, value)
             # The value should not be set
-            assert getattr(params, name) != value
+            __check_assigned_value_is_not_equal(getattr(params, name), value)
+    set_validation_mode(old_mode)
 
 
 if __name__ == "__main__":